Reordered CPR strategies
Always try to use the global decode method as it is able to check itself against even and odd frames, yet only on the dondition the samples are within 10 seconds.

Otherwise, try the local method that is seeded with the last emission. This only works if the transmission happened 10 seconds ago.

If all else fails, try the local method seeded by the user's position.